chore: move custom arg parsing to separate module for reuse
Technically this is a breaking change because of the Error type change, but not from CLI perspective
